Add script for Zoom meeting light control
Added a new AppleScript that automatically controls a light during Zoom meetings. The script checks for active Zoom meetings every second. When a meeting starts, it runs a shortcut named "Control Futton Light" to turn the light off and when the meeting ends, it runs the shortcut again to turn the light back on. This is for improving the visibility during online meetings.

(cherry picked from commit 76c2f2177fab1256e7b9c4828898fd4020319cb9)